Sorting Visualizer Project using Java
A simple Java Swing application that visualizes sorting algorithms (Bubble Sort, Selection Sort and Insertion Sort). Users can input an array of numbers, and the program visualizes the sorting process step-by-step. It includes options to change the array size and reset the visualization. This project demonstrates GUI development with Java Swing and algorithmic problem-solving for sorting.